third_party/blink/web_tests/platform/generic/editing/caret/selection-with-caret-type-progress-expected.txt

Tests whether explicitly set caret style is retained on performing text selection

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Try selecting this text by dragging the cursor. Progress cursor should be displayed while doing so.
PASS cursorInfo is "type=Progress"
PASS successfullyParsed is true

TEST COMPLETE
